requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autocommand==2.2.2
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.3
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.4.0
filelock==3.20.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.2
meson==1.9.2
more-itertools==10.8.0
packaging==25.0
pillow==12.0.0
pillow-avif-plugin==1.5.2
pillow_heif==1.1.1
platformdirs==4.5.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyGObject==3.54.5
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
replicate==1.0.7
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
virtualenv==20.35.4
wheel==0.45.1
opencv-python
numpy
gradio_client
requests
torch
torchvision
transformers
kornia
timm
